+++ Dockerfile
@@ -1,6 +1,6 @@
 FROM node:latest
 COPY . /src
-WORKDIR /src/src
+WORKDIR /src
 RUN ls
 RUN npm install
 CMD [ "npm", "start" ]
