+++ src/server.js
@@ -4,8 +4,8 @@ const formidable = require('formidable')
 const cors = require('cors')
 require('dotenv').config()
 const KeplerConfiguration = require('./KeplerConfiguration')
-const DataNotion = require('./src/helpers/notion')
-const DataWordpress = require('./src/helpers/wordpress')
+const DataNotion = require('./helpers/notion')
+const DataWordpress = require('./helpers/wordpress')
 const path = require('path')
 const fs = require('fs')
